Initial commit of geoips2_template_plugin repository
This provides an example plugin repository for additional functionality beyond what is available directly within the
geoips2 repository.

Complete instructions for installing and testing geoips2_template_repo included within this repo, as well as
instructions for modifying it for your own development purposes.

This repository simply duplicates and renames functionality that already exists within the geoips2 repo,
as a demonstration of how to setup a new plugin repository in order to implement new functionality.

This repository includes the following duplicated functionality:
1. algorithms/single_channel_test.py (geoips2 single_channel.py)
2. filename_formats/test_fname.py (geoips2 geoips_fname.py)
3. interpolation/pyresample_wrappers/interp_gauss_test.py (geoips2 interp_gauss.py)
4. output_formats/imagery_test.py (geoips2 imagery_clean.py)
5. user_colormaps/visir/Infrared_Test.py (geoips2 Infrared.py)
6. product_params/visir/Infrared-Test.yaml (geoips2 Infrared.yaml)