Clarify README.md for required updates with new plugins
Note that default_algorithm.py is an optional file and can be removed if no "helper" functions or modules
are required outside the interface_modules and yaml_configs directories.

Include LICENSE in list of files that potentially require modification.

Update required modifications to include tests/scripts/*.sh vs tests/scripts/data_types and output_formats.
    This was a deprecated test script organization, update to latest tests/scripts/*.sh

Update README_PLUGIN.md to note installation and test instructions in README.md, and required updates to README

Note README.md is one of the files that will need modified for a new plugin, and that
    README_PLUGIN.md is not a required file and can be removed.